% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/autocorrelationtest.R
\name{autocorrelation.test}
\alias{autocorrelation.test}
\title{Monte Carlo Test for Autocorrelation}
\usage{
autocorrelation.test(X, vectors = NULL, values = NULL, W = NULL,
  test = c("positive", "negative", "two-sided"), N = 999)
}
\arguments{
\item{X}{a vector of values}

\item{vectors}{an optional matrix of eigenvectors if no neighborhood matrix is supplied}

\item{values}{an optional vector Moran's I coefficients corresponding to each eigenvector supplied}

\item{W}{a neighborhood matrix}

\item{test}{a character sting specifying the alternative hypothesis, must be one of "positve" (default), "negative", or "two-sided".}

\item{N}{the number of permutations or randomizations; default is 999}
}
\description{
Performs a Monte Carlo test for autocorrelation using the method described by Dray (2011)
#'
}
\details{
When calculating Moran's I for a vector, the value observed represents a combination of both negative and positive autocorrelation forces. This can be problematic when negative autocorrelation forces are as stong as the positive autcorrelation forces. When this is the case, Moran's I might be low despite the presence of both positive and autocorrelation. Dray (2000) demonstrates that the positive and negative contributions to Moran's I can be calculated seperately, and this calculation can be used to have a more robust test for positive and/or negative autocorrelation.
}
\seealso{
\code{\link{weightmatrix}}, \code{\link{morans.i}}
}

