% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dbmem.R
\name{dbmem}
\alias{dbmem}
\title{Distance-based Moran's Eigenvector Maps}
\usage{
dbmem(dmat, X = NULL)
}
\arguments{
\item{X}{an optional covariate matrix to reduce multicolinearity between created eigenvectors and covariates}

\item{W}{a neighborhood matrix}
}
\value{
a list of (1) a matrix of eigenvectors, (2) eigenvalues (scaled to represent the respective Moran's I value for each eigenvector), and (3) the neighborhood matrix used (sometimes different than one supplied if asymmsetric).
}
\description{
Performs a distance-based Moran's Eigenvector Maps on a given neighborhood matrix (\strong{W})
}
\details{
This is a function to perform principal coordinates of neighbor matrices (PCNM) on a weighted neighborhood matrix (\strong{W}). By performing PCNM on this matrix, we are able to use the eigenvectors as regressors within a linear model context, thus filtering any dependent structures in our data by partitioning the autocorrelation into synthetic variates.

This filtering approach was first proposed by Griffith (2000) using a binary connectivity matrix, where the eigenvectors and eigenvalues are extracted from the matrix operation (\strong{I}\eqn{-}\strong{jj}'/\emph{n})\strong{C}(\strong{I}\eqn{-}\strong{jj}'/\emph{n}) where \emph{n} is the number of points in the domain, \strong{j} is  a vector of ones with length \emph{n}, \strong{I} is the identity matrix, and \strong{C} is a connectivity matrix.  Note that this operation can be performed is a connectivity matrix is supplied rather than an weighted neighborhood matrix.  As Griffith and Peres-Neto (2006) show, this is similar to the PCNM procedure proposed by Borcard and Legendre (2002; see below).

When this filtering method is employed correctly, the first extracted eigenvector (V1) is a set of values that has the greatest Moran's I coefficient achievable from the connectivity matrix, the second eigenvector represents the second greatest achievable, and so on...

Griffith (2000) suggested that these eigenvectors represent distinct pattern descriptions of latent autocorrelation in the variables.

The PCMN filtering method (Borcard and Legendre 2002) performs a similar matrix operation on a truncated weight matrix (\strong{W}) instead of a connectivity matrix (\strong{C}). Dray et al. (2006) further modified this approach and demonstrated that it falls under a class of multivariate approaches known as Moran's Eigenvector Maps. In particular the Borcard and Legendre (2002) approach is known as a distance-based Moran's Eigenvector Map (dbmem).

Also included in this function is the ability to supply a covariate matrix (\strong{X}). This is for the case where you might have multicolinearity issues between \strong{X} and the derived eigenvectors. Griffith (2000) demonstrates this can be performed with a connectivity matrix; however, it's validity for a weight matrix is not known.
}
\examples{
library(matrix)
set.seed(200)
n<-100
y<-rnorm(n)
D<-array(dim=rep(length(y),2))
diag(D)<-0
D[lower.tri(D,diag=F)]<-runif(length(D[lower.tri(D,diag=F)]),0,500)
library(Matrix)
D<-forceSymmetric(D,uplo="L")
W<-weighmatrix(D)
out<-dbmem(W) #Perform dbmem
out$values #See Moran's I values for each eigenvector
with(out,plot(vectors[,1],type="l")) #Visualize pattern described by eigenvector
}
\references{
Borcard, D., and P. Legendre. 2002. All-scale spatial analysis of ecological data by means of principal coordinates of neighbour matrices. Ecological Modelling 153: 51-68.

Dray, S., P. Legendre, and P.R. Peres-Neto. 2006. Spatial modeling: a comprehensive framework for principal coordinate analysis of neighbor matrices (PCMN). Ecological Modelling 196:483-493.

Griffith, D. 2000. A linear regression solution to the spatial autocorrelation problem. Journal of Geographical Systems 2: 141-156.
}
\seealso{
\code{\link{weightmatrix}}, \code{\link{morans.test}}
}

