% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/morancontribtest.R
\name{morancontrib.test}
\alias{morancontrib.test}
\title{Monte-Carlo Test for Positive or Negative Contributions to Moran's I}
\usage{
morancontrib.test(X, vectors = NULL, values = NULL, W = NULL,
  stat = c("positive", "negative"), N = 999)
}
\arguments{
\item{X}{a vector of values}

\item{vectors}{an optional matrix of eigenvectors if no neighborhood matrix is supplied}

\item{values}{an optional vector Moran's I coefficients corresponding to each eigenvector supplied}

\item{W}{a neighborhood matrix}

\item{stat}{a character sting specifying the whether "positve" (default) or "negative" contributions should be calculated.}

\item{N}{the number of permutations (default=999)}
}
\description{
Performs a randomization test for either postive or negative autocorrelation contributions to Moran's I using the method described by Dray (2011)
#'
}
\details{
When calculating Moran's I for a vector, the value observed represents a combination of both negative and positive autocorrelation forces. This can be problematic when negative autocorrelation forces are as stong as the positive autcorrelation forces. When this is the case, Moran's I might be low despite the presence of both positive and autocorrelation. Dray (2000) demonstrates that the positive and negative contributions to Moran's I can be calculated seperately. Using the formulation, we can test for signficant contribuitons to either positive or negative autocorrelation.
}

