% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/moransi.R
\name{morans.I}
\alias{morans.I}
\title{Calculate Moran's I for a vector}
\usage{
morans.I(X, W)
}
\arguments{
\item{X}{a vector of values}

\item{W}{a neighborhood matrix}
}
\value{
Moran's I
}
\description{
Calculates Moran's I with a given vector (\strong{X}) and a neighborhood matrix (\strong{W})
}
\details{
Moran's I is a value that represents the degree of autocorrelation in a vector (\strong{x}) given a weight matrix (\strong{W}). It is similar to calculating the correlation between two variables and is typically restricted to be between -1 and 1; however, this depends on whether the weight matrix is weighted by row. Negative values of Moran's I represent negative autocorrelation and positive values represent positive autocorrelation. Although these values can be compared to the spatial autocorrelation of other vectors, there is little utility of the Moran's I value itself.
}
\examples{
############
#Not run   #
############
library(matrix)
n<-100
y<-rnorm(n)
D<-array(dim=rep(length(y),2))
diag(D)<-0
D[lower.tri(D,diag=F)]<-runif(length(D[lower.tri(D,diag=F)]),0,500)
library(Matrix)
D<-forceSymmetric(D,uplo="L")
}
\seealso{
\code{\link{weightmatrix}}, \code{\link{morans.test}}
}

