% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/randomizevector.R
\name{randomize_vector}
\alias{randomize_vector}
\title{Randomize vector for Monte Carlo tests.}
\usage{
randomize_vector(X, N)
}
\arguments{
\item{X}{the vector to randomize}

\item{N}{the number of randomizations to be made}
}
\value{
A list of N vectors that represent randomizations of X
}
\description{
Creates a list of N randomizations of a vector (X)
}

