% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simautocor.R
\name{sim.autocorrelated}
\alias{sim.autocorrelated}
\title{Simulate Autocorrelated Data}
\usage{
sim.autocorrelated(X, rho, W)
}
\arguments{
\item{X}{a vector of values}

\item{rho}{stength of correlation}

\item{W}{a neighbohood matrix}
}
\description{
Simulates autocorrelated data given a vector (\strong{X}), \eqn{\rho}, and a neighborhood matrix (\strong{W})
}

