% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stcontrib.R
\name{stcontrib.test}
\alias{stcontrib.test}
\title{Tests for Spatiotemporal Contribution}
\usage{
stcontrib.test(Y, trans = c("chord", "hellinger"), stat = c("TCV", "SCV"),
  N = 999, ...)
}
\arguments{
\item{Y}{a time by site matrix of abudances}

\item{trans}{a string determining the transformation used on Y; currently implemented options are: 'chord' (default) and 'hellinger'}

\item{stat}{a string determining whether to test for temporal (TCV; default) or spatial contributions (SCV) to variance}

\item{N}{the number of randomizations to perform}
}
\value{
a vector of emperical p-values for either temporal (TCV) or spatial (SCV) contributions to variance
}
\description{
Tests for significant temporal (TCV) and spatial (SCV) contributions to variance
}
\details{
This function is used to test for signficant temporal (TCV) or spatial (SCV) contributions to spatiotemporal variance in abundance or counts of a species.
}
\seealso{
\code{\link{stvariance}}
}

