% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stvariance.R
\name{stvariance}
\alias{stvariance}
\title{Calculate Spatiotemporal Variation}
\usage{
stvariance(Y, trans = c("chord", "hellinger"), print.results = T, ...)
}
\arguments{
\item{Y}{a time by site matrix of abudances}

\item{trans}{a string determining the transformation used on Y; currently implemented options are: 'chord' (default) and 'hellinger'}

\item{print.results}{a logical result to print results (default) or be silent}
}
\value{
a list including (1) total spatiotemporal variance, (2) temporal contributions to variance, and (3) spatial contributions to variance
}
\description{
Calculates and partitions spatiotemporal variation given a n (time) x p (site) matrix
}
\details{
This function is used to partition spatiotemporal variance in abundance into: (1) total spatiotemporal variance, (2) temporal contributions to variance, and (3) spatial contributions to variance.
}
\examples{
Create Data
Y<-array(c(0,2,1,5,0,0,0,3,1,2,4,2,0,0,2,2,0,2,2,0,0,2,3,1,5,8,4,2,1,2,2,2,4,3,1,2,4,4,3,2),dim=c(8,5))
rownames(Y)<-c(2006:2012,2014)
colnames(Y)<-1:5
}

