% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/weightmatrix.R
\name{weightmatrix}
\alias{weightmatrix}
\title{Create a weight matrix}
\usage{
weightmatrix(D, method = c("dbmem", "linear", "concave-up", "concave-down",
  "connectivity"), thresh = NULL, alpha = NULL, beta = NULL,
  diag.mat = 0)
}
\arguments{
\item{D}{a distance matrix or n x n matrix of distances to be used}

\item{method}{the method to be used to create the wegith matrix; must be one of "dbmem" (default), "linear", "concave-up", "concave-down", or "connectivity".}

\item{thresh}{a threshold value to be used in which sites are no longer connected; the default is the maximum distance on a minimum spanning tree.}

\item{alpha}{the power used for the "concave-down" method.}

\item{beta}{the power used for the "concave-up" method.}

\item{diag.mat}{a 0 or 1 to be used as the diagonal weight in the output matrix.}
}
\value{
A weighted neighboorhod matrix.
}
\description{
Creates a weighted neighborhood matrix from a distance matrix using different rule sets.
}
\details{
There are a variety of rules that one can use to define a neighborhood matrix, especially with lattice data (i.e., data located in regions). However, often we deal with data within a spatial (or temporal) domain that does not explicitly fit into regions, but rather they are points in space, time, or from an  evolutionary spectrum. For these cases, we need some way of defining whether a site, time, or species is a neighbor or not with another and a way of expressing the difficulty of exchange between other sites, times, or species.  This is why we create a weighted neighborhood matrix.

 Using a distance matrix, we can use a variety of rules to calculate a weighted neighborhood matrix.  This function uses four different methods that describe the increase in difficultly or exchange between points or species.  These functions are the ones described by Drey et al. (2006) and include: distance-based Moran's Eigenvector Maps (dbmem), Linear Fuction, Concave-Down, Concave-Up, and Connectivity.

 For the first four methods we need to define both (1) a connectivity matrix (B) and (2) an edge weight matrix (A); for the latter we only need a connectivity matrix.   To create a connectivity matrix, we need a thresdhold in which to say sites are no longer connected. The weightmatrix function allows for the user to provide a value; however, there is a standard theshold used for dbmem.  This is usually difined as the maximum distance on a minimum span treee.   This span tree is simply a tree that connects all sites together with edges, and the minimum span tree is the tree with the lowest cost. With the connectivity matrix defined, an edge matrix can be calculated using the different rule sets.

 To compute the weighted neighborhood we just need to take the Hadamard product of B and A.  However, the diagonal will be ones representing a connection.  This is good if one is going to try to derive more complex relationships with other weight matrices (e.g., spatiotemporal designs), but we might want to have zeros on the diagonal instead (default of function).  This is controled by the diag.mat argument.
}
\examples{
#Not run
 D<-as.matrix(dist(cbind(runif(100,0,100),runif(100,0,100))))
 #Distance-based Moran's Eigenvector Maps
 W<-weightmatrix(D,method='dbmem')
 W

 #Concave-Up
 W<-weightmatrix(D,method='concave-up',beta=2)
 W
}

