% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stcontrib.R
\name{stcontrib.test}
\alias{stcontrib.test}
\title{Tests for Spatiotemporal Contribution}
\usage{
stcontrib.test(Y, trans = c("chord", "hellinger"), stat = c("TCV", "SCV"),
  N = 999, ...)
}
\arguments{
\item{Y}{a time by site matrix of abudances}

\item{trans}{a string determining the transformation used on Y; currently implemented options are: 'chord' (default) and 'hellinger'}

\item{stat}{a string determining whether to test for temporal (TCV; default) or spatial contributions (SCV) to variance}

\item{N}{the number of randomizations to perform}
}
\value{
a vector of emperical p-values for either temporal (TCV) or spatial (SCV) contributions to variance
}
\description{
Tests for significant temporal (TCV) and spatial (SCV) contributions to variance
}
\details{
This function is used to test for signficant temporal (TCV) or spatial (SCV) contributions to spatiotemporal variance in abundance or counts of a species.
}
\examples{
Load data from Percina rex Smith River sites
Y<-array(c(0,2,1,5,0,0,0,3,
1,2,4,2,0,0,2,2,0,2,2,0,0,2,
3,1,5,8,4,2,1,2,2,2,4,3,1,2,
4,4,3,2),dim=c(8,5))
rownames(Y)<-c(2006:2012,2014)
colnames(Y)<-1:5
stcontrib.test(Y,stat="TCV") #Test for signficant temporal contributions
stcontrib.test(Y,stat="SCV") #Test for significant spatial contriubitons
}
\seealso{
\code{\link{stvariance}}
}

